initial dirty version of action plugin for node add/remove
This action plugin allows to determine what action should be done
when desired node_list differs from actual cluster node list that
is detected (TODO fact collection plugin not developed yet, use
of attributes in inventory needed for testing)

When this action plugin knows what needs to be done it will call
the pcs_cluster module with possibly adjusted attributes on nodes
where it determines it is needed.

Removal of node from cluster is done by node itself.
Add of node to cluster is done first node in existing cluster.

It is not possible to add and remove nodes in single action.