dirty version 2 of the "add/remove node"-capable pcs_cluster module
This actually works for simple use cases and for some more complicated
ones. Most importantly this has a fact_module for detecting EL7
pacemaker cluster on node and parsing the node names out of it.

Then the action plugin for pcs_cluster decides which node should do what
and the actual pcs_cluster module does the dirty job.

This should be able to add and remove nodes from existing clusters,
create a new cluster, create a new cluster on nodes where is no cluster
even if in play there are other nodes with existing cluster.

TODO: make the logic in action plugin more clear on what is
allowed/possible to do with this.